# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FreightAudit API — análise, auditoria e simulação de remuneração de fretes
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /snapshots:
    get:
      operationId: listSnapshots
      tags: [snapshots]
      summary: Listar todos os snapshots
      responses:
        "200":
          description: Lista de snapshots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snapshot"
    post:
      operationId: createSnapshot
      tags: [snapshots]
      summary: Criar snapshot manualmente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotInput"
      responses:
        "201":
          description: Snapshot criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
  /snapshots/current:
    get:
      operationId: getCurrentSnapshot
      tags: [snapshots]
      summary: Obter snapshot atual (is_current = true)
      responses:
        "200":
          description: Snapshot atual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "404":
          description: Nenhum snapshot atual configurado
  /snapshots/{id}:
    get:
      operationId: getSnapshot
      tags: [snapshots]
      summary: Obter snapshot por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Snapshot encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "404":
          description: Snapshot não encontrado
    patch:
      operationId: updateSnapshot
      tags: [snapshots]
      summary: Atualizar snapshot (label, status, is_current)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnapshotUpdate"
      responses:
        "200":
          description: Snapshot atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snapshot"
        "404":
          description: Snapshot não encontrado
  /snapshots/{id}/parameters:
    get:
      operationId: getSnapshotParameters
      tags: [snapshots]
      summary: Listar parâmetros de um snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de parâmetros do snapshot
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Parameter"

  # ─── PARAMETERS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Snapshot:
      type: object
      required: [id, label, snapshotDate, status, isCurrent, parameterCount, createdAt]
      properties:
        id:
          type: string
        label:
          type: string
        snapshotDate:
          type: string
          format: date
        importJobId:
          type: ["string", "null"]
        status:
          type: string
          enum: [DRAFT, ACTIVE, ARCHIVED]
        isCurrent:
          type: boolean
        parameterCount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    SnapshotInput:
      type: object
      required: [label, snapshotDate]
      properties:
        label:
          type: string
          minLength: 1
        snapshotDate:
          type: string
          format: date
        notes:
          type: string
    SnapshotUpdate:
      type: object
      properties:
        label:
          type: string
          minLength: 1
        status:
          type: string
          enum: [DRAFT, ACTIVE, ARCHIVED]
        isCurrent:
          type: boolean
        notes:
          type: string

    # ─── PARAMETER ──────────────────────────────────────────────────────────────
    Parameter:
      type: object
      required:
        - id
        - snapshotId
        - parameterKey
        - parameterName
        - category
        - dataType
        - effectiveFrom
        - version
        - isSuperseded
        - createdAt
      properties:
        id:
          type: string
        snapshotId:
          type: string
        parameterKey:
          type: string
        parameterName:
          type: string
        category:
          type: string
        dataType:
          type: string
        unit:
          type: ["string", "null"]
        scopeLabel:
          type: ["string", "null"]
        valueNumeric:
          type: ["number", "null"]
        valueText:
          type: ["string", "null"]
        effectiveFrom:
          type: string
          format: date
        effectiveUntil:
          type: ["string", "null"]
        version:
          type: number
        isSuperseded:
          type: boolean
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...